move text to speech out as a separate plugin
