+++ webpack.config.js
@@ -43,6 +43,9 @@ module.exports = {
 			'./src/js/media-modal/views/generate-image-media-upload.js',
 		],
 		'extend-image-blocks': './src/js/extend-image-block-generate-image.js',
+
+		'classifai-plugin-fill': './src/js/plugins/slot-fill/index.js',
+		'classifai-plugin-text-to-speech': './src/js/plugins/text-to-speech/index.js',
 	},
 	module: {
 		rules: [
